Fix circular import and cors issue in testing
